+++ R/barplot.R
@@ -97,13 +97,24 @@ stacked_barplot <- function(obj, tax_level = "Phylum", fill = "Phylum", xlabel =
   palette_values <- as.character(palette_values)
 
   # Create the theme
-  p <- p + ggplot2::ylab("Relative Abundance (% 100)") + ggplot2::scale_fill_manual(values = palette_values) + ggplot2::theme(
-    text = ggplot2::element_text(size = 11, face = "bold"),
-    axis.text.x = ggplot2::element_blank(), axis.ticks = ggplot2::element_blank(), panel.grid.major = ggplot2::element_blank(), panel.grid.minor = ggplot2::element_blank(),
-    strip.background = ggplot2::element_rect(fill = "white"), strip.text = ggplot2::element_text(colour = "black"), panel.background = ggplot2::element_blank()
-  ) +
+  p <- p +
+    ggplot2::ylab("Relative Abundance (% 100)") +
+    ggplot2::scale_fill_manual(values = palette_values) +
+    ggplot2::theme(
+      text = ggplot2::element_text(size = 11, face = "bold"),
+      axis.text.x = ggplot2::element_text(angle = 90, hjust = 1, vjust = 0.5),
+      axis.ticks = ggplot2::element_blank(),
+      panel.grid.major = ggplot2::element_blank(),
+      panel.grid.minor = ggplot2::element_blank(),
+      strip.background = ggplot2::element_rect(fill = "white"),
+      strip.text = ggplot2::element_text(colour = "black"),
+      panel.background = ggplot2::element_blank()
+    ) +
     ggplot2::xlab(xlabel) +
-    ggplot2::scale_y_continuous(breaks = scales::pretty_breaks(n = 4))
+    ggplot2::scale_y_continuous(
+      breaks = scales::pretty_breaks(n = 4),
+      expand = ggplot2::expansion(mult = c(0, 0))
+    )
 
   # Add faceting, if true
   if (faceted) {